Disk: allow setting the default partitioning mode of a PT
Some image types want to default to a different partitioning mode than
the AutoLVM. An example of such image types are the EDGE images on
RHEL-8.

Extend the `disk.PartitionTable` struct to allow setting the default
partitioning mode of a PT. If not set, the old behavior is preserved
(AutoLVM).

Cover the functionality with a unit test.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>