Test: add testing manifest for RHEL-33568
RHEL-33568 is an issue which causes depsolve errors when a user uses
customization which adds a package to the base package set and at the
same time also specifies the same package in the BP, but with lower than
the latest version in the repositories.

Base the testing manifest on F39. Repos that we use contain:
chrony-0:4.4-1.fc39.x86_64
chrony-0:4.5-1.fc39.x86_64
scap-security-guide-0:0.1.68-1.fc39.noarch
scap-security-guide-0:0.1.72-1.fc39.noarch

The test manifest uses NTP servers and OSCAP customization, while
requesting the older versions of chrony and scap-security-guide
packages.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>