pkg: add 'sbom' package for working with SBOM documents
Add a new `sbom` package for working with SBOM documents. It provides a
simple wrapper struct, which currently supports only SPDX standard. The
`sbom.Document` can be created from SPDX JSON SBOM and then serialized
back to SPDX JSON SBOM document. One can however it also as an
`spdx.Document` instance, which can be serialized to any supported SPDX
output format.

The idea is to provide a bit of an abstraction from the specific SBOM
implementation, so that in the future, it would be possible to create
`sbom.Document` from SBOM documents of various standards and also
serialize it back to various SBOM standards.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>