test/scripts/dl-image-build-cache: generate manifests in a container
In order for the script to download the image build cache from S3, it
needs to generate test manifests to determine their checksum. It turns
out that this may be a problem on older or currently GA RHEL releases,
which may be shipping lower Go compiler version, than the minimum
required version by the images go.mod.

To workaround this situation on specific distributions, allow
generating test manifests in a container. This will allow generating
test manifests on RHEL using Fedora container image.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>